+++ dotnet/src/webdriver/BiDi/BrowsingContext/ContextDestroyedEvent.cs
@@ -1,4 +1,4 @@
-// <copyright file="FileDialogOpenedEventArgs.cs" company="Selenium Committers">
+// <copyright file="ContextDestroyedEvent.cs" company="Selenium Committers">
 // Licensed to the Software Freedom Conservancy (SFC) under one
 // or more contributor license agreements.  See the NOTICE file
 // distributed with this work for additional information
@@ -17,18 +17,12 @@
 // under the License.
 // </copyright>
 
-namespace OpenQA.Selenium.BiDi.Input;
+namespace OpenQA.Selenium.BiDi.BrowsingContext;
 
-public sealed record FileDialogOpenedEventArgs(
-    IBiDi BiDi,
-    BrowsingContext.BrowsingContext Context,
-    Browser.UserContext? UserContext,
-    bool Multiple,
-    Script.SharedReference? Element)
-    : EventArgs(BiDi);
-
-internal sealed record FileDialogInfo(
-    BrowsingContext.BrowsingContext Context,
-    Browser.UserContext? UserContext,
-    bool Multiple,
-    Script.SharedReference? Element);
+public sealed record ContextDestroyedEventArgs(IBiDi BiDi, IReadOnlyList<Info>? Children,
+    Browser.ClientWindow ClientWindow,
+    BrowsingContext Context,
+    BrowsingContext? OriginalOpener,
+    string Url,
+    Browser.UserContext UserContext,
+    BrowsingContext? Parent) : EventArgs(BiDi);
