+++ dotnet/src/webdriver/BiDi/BrowsingContext/LoadEvent.cs
@@ -1,4 +1,4 @@
-// <copyright file="NavigationCommittedEventArgs.cs" company="Selenium Committers">
+// <copyright file="LoadEvent.cs" company="Selenium Committers">
 // Licensed to the Software Freedom Conservancy (SFC) under one
 // or more contributor license agreements.  See the NOTICE file
 // distributed with this work for additional information
@@ -19,5 +19,5 @@
 
 namespace OpenQA.Selenium.BiDi.BrowsingContext;
 
-public sealed record NavigationCommittedEventArgs(IBiDi BiDi, BrowsingContext Context, Navigation? Navigation, DateTimeOffset Timestamp, string Url, Browser.UserContext? UserContext)
+public sealed record LoadEventArgs(IBiDi BiDi, BrowsingContext Context, Navigation? Navigation, DateTimeOffset Timestamp, string Url, Browser.UserContext? UserContext)
     : NavigationEventArgs(BiDi, Context, Navigation, Timestamp, Url, UserContext);
