+++ dotnet/src/webdriver/BiDi/BrowsingContext/NavigationFailedEvent.cs
@@ -1,4 +1,4 @@
-// <copyright file="RealmDestroyedEventArgs.cs" company="Selenium Committers">
+// <copyright file="NavigationFailedEvent.cs" company="Selenium Committers">
 // Licensed to the Software Freedom Conservancy (SFC) under one
 // or more contributor license agreements.  See the NOTICE file
 // distributed with this work for additional information
@@ -17,10 +17,7 @@
 // under the License.
 // </copyright>
 
-namespace OpenQA.Selenium.BiDi.Script;
+namespace OpenQA.Selenium.BiDi.BrowsingContext;
 
-public sealed record RealmDestroyedEventArgs(
-    IBiDi BiDi,
-    Realm Realm) : EventArgs(BiDi);
-
-internal sealed record RealmDestroyedParameters(Realm Realm);
+public sealed record NavigationFailedEventArgs(IBiDi BiDi, BrowsingContext Context, Navigation? Navigation, DateTimeOffset Timestamp, string Url, Browser.UserContext? UserContext)
+    : NavigationEventArgs(BiDi, Context, Navigation, Timestamp, Url, UserContext);
