+++ dotnet/src/webdriver/BiDi/Input/FileDialogOpenedEvent.cs
@@ -1,4 +1,4 @@
-// <copyright file="NavigationFailedEventArgs.cs" company="Selenium Committers">
+// <copyright file="FileDialogOpenedEvent.cs" company="Selenium Committers">
 // Licensed to the Software Freedom Conservancy (SFC) under one
 // or more contributor license agreements.  See the NOTICE file
 // distributed with this work for additional information
@@ -17,7 +17,18 @@
 // under the License.
 // </copyright>
 
-namespace OpenQA.Selenium.BiDi.BrowsingContext;
+namespace OpenQA.Selenium.BiDi.Input;
 
-public sealed record NavigationFailedEventArgs(IBiDi BiDi, BrowsingContext Context, Navigation? Navigation, DateTimeOffset Timestamp, string Url, Browser.UserContext? UserContext)
-    : NavigationEventArgs(BiDi, Context, Navigation, Timestamp, Url, UserContext);
+public sealed record FileDialogOpenedEventArgs(
+    IBiDi BiDi,
+    BrowsingContext.BrowsingContext Context,
+    Browser.UserContext? UserContext,
+    bool Multiple,
+    Script.SharedReference? Element)
+    : EventArgs(BiDi);
+
+internal sealed record FileDialogInfo(
+    BrowsingContext.BrowsingContext Context,
+    Browser.UserContext? UserContext,
+    bool Multiple,
+    Script.SharedReference? Element);
