+++ dotnet/src/webdriver/BiDi/Log/EntryAddedEvent.cs
@@ -1,4 +1,4 @@
-// <copyright file="EntryAddedEventArgs.cs" company="Selenium Committers">
+// <copyright file="EntryAddedEvent.cs" company="Selenium Committers">
 // Licensed to the Software Freedom Conservancy (SFC) under one
 // or more contributor license agreements.  See the NOTICE file
 // distributed with this work for additional information
