+++ dotnet/src/webdriver/BiDi/Script/MessageEvent.cs
@@ -1,4 +1,4 @@
-// <copyright file="FragmentNavigatedEventArgs.cs" company="Selenium Committers">
+// <copyright file="MessageEvent.cs" company="Selenium Committers">
 // Licensed to the Software Freedom Conservancy (SFC) under one
 // or more contributor license agreements.  See the NOTICE file
 // distributed with this work for additional information
@@ -17,7 +17,15 @@
 // under the License.
 // </copyright>
 
-namespace OpenQA.Selenium.BiDi.BrowsingContext;
+namespace OpenQA.Selenium.BiDi.Script;
 
-public sealed record FragmentNavigatedEventArgs(IBiDi BiDi, BrowsingContext Context, Navigation? Navigation, DateTimeOffset Timestamp, string Url, Browser.UserContext? UserContext)
-    : NavigationEventArgs(BiDi, Context, Navigation, Timestamp, Url, UserContext);
+public sealed record MessageEventArgs(
+    IBiDi BiDi,
+    Channel Channel,
+    RemoteValue Data,
+    Source Source) : EventArgs(BiDi);
+
+internal sealed record MessageParameters(
+    Channel Channel,
+    RemoteValue Data,
+    Source Source);
